srcBDD/mySecondAPI_with_MVC/alchimie/requirements.txt
flask
flask_sqlalchemy
pyjwt
flask-swagger-ui